@@ -15,12 +15,11 @@
     }
   ],
   "require": {
-    "php": ">=7.3",
-    "codeclimate/php-test-reporter": "^0.2.0"
+    "php": ">=8.3"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9",
-    "symfony/console": "~2.5 || ~3.0 || ~4.0"
+    "phpunit/phpunit": "*",
+    "symfony/console": "*"
   },
   "autoload": {
     "psr-4": {
